pyproject.toml
[project]
name = "data-creating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.15.0",
    "pandas>=2.2.3",
]
